Create a helper file to calculate the meta description length
