+++ packages/yoastseo/src/languageProcessing/index.js
@@ -23,7 +23,7 @@ import areWordsInSentence from "./helpers/word/areWordsInSentence";
 import getClauses from "./helpers/passiveVoice/periphrastic/getClauses";
 import getClausesSplitOnStopWords from "./helpers/passiveVoice/periphrastic/freeAuxiliaryParticipleOrder/getClausesSplitOnStopWords";
 import { stripFullTags as stripHTMLTags } from "./helpers/sanitize/stripHTMLTags";
-import metaDescriptionLength from "./researches/metaDescriptionLength";
+import countMetaDescriptionLength from "./helpers/word/countMetaDescriptionLength";
 
 export {
 	AbstractResearcher,
@@ -52,5 +52,5 @@ export {
 	getClauses,
 	getClausesSplitOnStopWords,
 	stripHTMLTags,
-	metaDescriptionLength,
+	countMetaDescriptionLength,
 };
